[FIX] Odoo 9-10: `reload` handling
We were still trying to make the auto reload feature work with the 8.0
way in the dockerfile 9 and 10. This patch install the dependency used
in Odoo 9-10 (watchdog instead of pyinotify) and remove the
"auto-reload" key from the default configuration file, as the new syntax
is "--dev=reload".

The auto-reload is now disabled by default (it didn't work, let's not
bother someone that does not need it). The easiest way to use it is to
pass the arguments to the odoo process as:

docker run -p 8069:8069 --name odoo --link db:db -t odoo:[9.0|10.0] -- --dev=reload

fixes #84